+++ cli/build_worker.ts
@@ -1,6 +1,6 @@
 import { log } from "../core/utils/log.ts";
 import { localIp, openBrowser } from "../core/utils/net.ts";
-import { setEnv } from "../core/utils/env.ts";
+import { env, setEnv } from "../core/utils/env.ts";
 import { normalizePath } from "../core/utils/path.ts";
 import { resolveConfigFile } from "../core/utils/lume_config.ts";
 import { fromFileUrl } from "../deps/path.ts";
@@ -35,6 +35,12 @@ interface BuildOptions {
 async function build({ type, config, serve, cms: loadCms }: BuildOptions) {
   // Set the live reload environment variable to add hash to the URLs in the module loader
   setEnv("LUME_LIVE_RELOAD", "true");
+
+  // Show draft pages in development mode (if not set already)
+  const showDrafts = env<boolean | undefined>("LUME_DRAFTS");
+  if (showDrafts === undefined) {
+    setEnv("LUME_DRAFTS", "true");
+  }
 
   const _config = await resolveConfigFile(["_config.ts", "_config.js"], config);
   const site = await createSite(_config);